+++ src/chainlit/frontend/src/components/tasklist/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './TaskList';
