+++ src/chainlit/frontend/src/components/tasklist/types.d.ts
@@ -1,0 +1,9 @@
+export interface ITask {
+  title: string;
+  status: 'ready' | 'running' | 'done' | 'failed';
+}
+
+export interface ITaskList {
+  status: 'ready' | 'running' | 'done';
+  tasks: ITask[];
+}
